Move KV placement helpers out of the MHA layout
